+++ src/main/java/org/jboss/xavier/integrations/route/model/cloudforms/v1/ManageIQProvidersVmwareInfraManager.java
@@ -26,7 +26,7 @@ import java.util.Map;
 public class ManageIQProvidersVmwareInfraManager {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
     @JsonProperty("name")
     private String name;
     @JsonProperty("type")
@@ -45,16 +45,16 @@ public class ManageIQProvidersVmwareInfraManager {
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("id")
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     @JsonProperty("id")
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public ManageIQProvidersVmwareInfraManager withId(Integer id) {
+    public ManageIQProvidersVmwareInfraManager withId(Long id) {
         this.id = id;
         return this;
     }
